Galutiniai pakeitimai

+++ controllers/samples.js
@@ -4,6 +4,7 @@ const { cloudinary } = require('../cloudinary')
 
 
 
+
 module.exports.index = async (req, res) => {
     const samples = await Sample.find({}).populate('author')
     res.render('samples', { samples })
@@ -14,32 +15,63 @@ module.exports.newForm = (req, res) => {
 }
 
 module.exports.new =  async (req, res) => {
-    const { sampleName, type, key } = req.body
-    const url = req.file.path
-    const filename = req.file.filename
-    const user = await User.findById(req.user.id)
-    const sample = new Sample({ name: sampleName, filename: filename, sampleUrl: url, type: type, key: key, author: user })
-    await sample.save()
-    res.redirect('/samples')
+    try{
+        const { sampleName, type, key } = req.body
+        const url = req.file.path
+        const filename = req.file.filename
+        const user = await User.findById(req.user.id)
+        const sample = new Sample({ name: sampleName, filename: filename, sampleUrl: url, type: type, key: key, author: user })
+        await sample.save()
+        req.flash('success', 'Sample added successfully')
+        res.redirect('/samples')
+    }catch(e){
+        req.flash('error', 'Something went wrong')
+        return res.redirect(`/samples`)
+    }
 }
 
 module.exports.delete = async (req, res) => {
-    const { id } = req.params;
-    const sample = await Sample.findById(id);
-    await cloudinary.uploader.destroy(sample.filename)
-    await Sample.findByIdAndRemove(id);
-    res.redirect('/samples')
+    try{
+        const { id } = req.params;
+        const sample = await Sample.findById(id);
+        if (!sample) {
+            req.flash('error', 'Cannot find this blog')
+            return res.redirect(`/samples`)
+        }
+        await cloudinary.uploader.destroy(sample.filename, {"resource_type": "video"})
+        await Sample.findByIdAndRemove(id);
+        req.flash('success', 'Sample deleted successfully')
+        res.redirect('/samples')
+    }catch(e){
+        req.flash('error', 'Something went wrong')
+        return res.redirect(`/samples`)
+    }
 }
 
 module.exports.editForm = async (req, res) => {
-    const { id } = req.params;
-    const sample = await Sample.findById(id);
-    res.render('samples/edit', {sample})
+    try{
+        const { id } = req.params;
+        const sample = await Sample.findById(id);
+        if (!sample) {
+            req.flash('error', 'Cannot find this sample')
+            return res.redirect(`/samples`)
+        }
+        res.render('samples/edit', {sample})
+    }catch(e){
+        req.flash('error', 'Something went wrong')
+        return res.redirect(`/samples`)
+    }
 }
 
 module.exports.edit = async (req, res) => {
-    const{sampleName, type, key} = req.body
-    const { id } = req.params;
-    await Sample.findByIdAndUpdate(id, {name:sampleName, type:type, key:key});
-    res.redirect('/samples')
+    try{
+        const{sampleName, type, key} = req.body
+        const { id } = req.params;
+        await Sample.findByIdAndUpdate(id, {name:sampleName, type:type, key:key});
+        req.flash('success', 'Sample updated successfully')
+        res.redirect('/samples')
+    }catch(e){
+        req.flash('error', 'Something went wrong')
+        return res.redirect(`/samples`)
+    }
 }
